# Conflicts:
#	concrete/single_pages/dashboard/system/environment/entities.php
#	tests/tests/Core/Package/fixtures/packages/test_metadatadriver_xml/controller.php
#	tests/tests/Core/Package/fixtures/packages/test_metadatadriver_yaml/controller.php


Former-commit-id: bfd393b6041e43fe609f3f483464519553ac4569
Former-commit-id: 14a8e24d0a8cdfeec23bcea3a919ff8aee531647

+++ concrete/src/Page/Type/Composer/Control/BlockControl.php
@@ -2,8 +2,6 @@
 namespace Concrete\Core\Page\Type\Composer\Control;
 
 use Concrete\Core\Package\PackageList;
-use Loader;
-use Concrete\Core\Foundation\Object;
 use Controller;
 use Package;
 use Block;
@@ -13,6 +11,8 @@ use Concrete\Core\Page\Page;
 use Area;
 use PageTemplate;
 use Concrete\Core\Page\Type\Composer\FormLayoutSet as PageTypeComposerFormLayoutSet;
+use Concrete\Core\Support\Facade\Application;
+use Concrete\Core\View\View;
 
 class BlockControl extends Control
 {
@@ -50,36 +50,42 @@ class BlockControl extends Control
         $b->deleteBlock();
     }
 
+    public function onPageDraftCreate(Page $c)
+    {
+        return $this->publishToPage($c, [], []);
+    }
+
     public function getPageTypeComposerControlBlockObject(Page $c)
     {
-        $db = Loader::db();
+        $app = Application::getFacadeApplication();
+        $db = $app->make('database')->connection();
         if (!is_object($this->b)) {
             $setControl = $this->getPageTypeComposerFormLayoutSetControlObject();
-            $r = $db->GetRow(
+            $r = $db->fetchAssoc(
                 $q = 'select cdb.bID, cdb.arHandle from PageTypeComposerOutputBlocks cdb inner join CollectionVersionBlocks cvb on (cdb.bID = cvb.bID and cvb.cID = cdb.cID and cvb.cvID = ?) where cdb.ptComposerFormLayoutSetControlID = ? and cdb.cID = ?',
-                array(
+                [
                     $c->getVersionID(),
                     $setControl->getPageTypeComposerFormLayoutSetControlID(),
                     $c->getCollectionID(),
-                )
+                ]
             );
-            if (!$r['bID']) {
+            if (!$r) {
                 // this is the first run. so we look for the proxy block.
                 $pt = PageTemplate::getByID($c->getPageTemplateID());
                 $outputControl = $setControl->getPageTypeComposerOutputControlObject($pt);
                 if (is_object($outputControl)) {
                     $cm = $c->getPageTypeObject();
                     $mc = $cm->getPageTypePageTemplateDefaultPageObject($pt);
-                    $r = $db->GetRow(
+                    $r = $db->fetchAssoc(
                         'select bco.bID, cvb.arHandle from btCorePageTypeComposerControlOutput bco inner join CollectionVersionBlocks cvb on cvb.bID = bco.bID where ptComposerOutputControlID = ? and cvb.cID = ?',
-                        array(
+                        [
                             $outputControl->getPageTypeComposerOutputControlID(),
                             $mc->getCollectionID(),
-                        )
+                        ]
                     );
                 }
             }
-            if ($r['bID']) {
+            if ($r) {
                 $b = Block::getByID($r['bID'], $c, $r['arHandle']);
                 $this->setPageTypeComposerControlBlockObject($b);
 
@@ -161,8 +167,9 @@ class BlockControl extends Control
     public function getPageTypeComposerControlCustomTemplates()
     {
         $bt = $this->getBlockTypeObject();
-        $txt = Loader::helper('text');
-        $templates = array();
+        $app = Application::getFacadeApplication();
+        $txt = $app->make('helper/text');
+        $templates = [];
         if (is_object($bt)) {
             $blocktemplates = $bt->getBlockTypeComposerTemplates();
             if (is_array($blocktemplates)) {
@@ -206,7 +213,7 @@ class BlockControl extends Control
             if ($this->page) {
                 // we HAVE a page, but we don't have a block object, which means something has gone wrong.
                 // we've lost the association. So we abort.
-                Loader::element('page_types/composer/controls/invalid_block');
+                View::element('page_types/composer/controls/invalid_block');
 
                 return;
             }
@@ -214,9 +221,9 @@ class BlockControl extends Control
         }
 
         $this->getController($obj);
-
+        $app = Application::getFacadeApplication();
         $env = Environment::get();
-        $form = Loader::helper('form');
+        $form = $app->make('helper/form');
         $set = $this->getPageTypeComposerFormLayoutSetControlObject()->getPageTypeComposerFormLayoutSetObject();
 
         if ($customTemplate) {
@@ -236,7 +243,6 @@ class BlockControl extends Control
                         $template = DIRNAME_BLOCK_TEMPLATES_COMPOSER . '/' . $customTemplate;
                         break;
                     }
-
                 }
             }
         }
@@ -245,7 +251,7 @@ class BlockControl extends Control
             $template = FILENAME_BLOCK_COMPOSER;
         }
 
-        $this->inc($template, array('control' => $this, 'obj' => $obj, 'description' => $description));
+        $this->inc($template, ['control' => $this, 'obj' => $obj, 'description' => $description]);
     }
 
     public function action($task)
@@ -253,27 +259,28 @@ class BlockControl extends Control
         $obj = $this->getPageTypeComposerControlDraftValue();
         if (!is_object($obj)) {
             // we don't have a page, an area, or ANYTHING YET.
-            $arguments = array('/ccm/system/block/action/add_composer',
+            $arguments = ['/ccm/system/block/action/add_composer',
                 $this->getPageTypeComposerFormLayoutSetControlObject()->getPageTypeComposerFormLayoutSetControlID(),
                 $task,
-            );
+            ];
 
-            return call_user_func_array(array('\URL', 'to'), $arguments);
+            return call_user_func_array(['\URL', 'to'], $arguments);
         } else {
             $area = $obj->getBlockAreaObject();
             $c = $area->getAreaCollectionObject();
-            $arguments = array('/ccm/system/block/action/edit_composer',
+            $arguments = [
+                '/ccm/system/block/action/edit_composer',
                 $c->getCollectionID(),
                 urlencode($area->getAreaHandle()),
                 $this->getPageTypeComposerFormLayoutSetControlObject()->getPageTypeComposerFormLayoutSetControlID(),
                 $task,
-            );
+            ];
 
-            return call_user_func_array(array('\URL', 'to'), $arguments);
+            return call_user_func_array(['\URL', 'to'], $arguments);
         }
     }
 
-    public function inc($file, $args = array())
+    public function inc($file, $args = [])
     {
         extract($args);
         if (!isset($obj)) {
@@ -309,9 +316,7 @@ class BlockControl extends Control
                 if (file_exists($include)) {
                     include $include;
                 }
-
             }
-
         }
     }
 
@@ -336,14 +341,15 @@ class BlockControl extends Control
 
         // delete the block that this set control has placed on this version, because
         // we are going to replace it with a new one.
-        $db = Loader::db();
+        $app = Application::getFacadeApplication();
+        $db = $app->make('database')->connection();
         $q = 'select cvb.arHandle, cdb.bID, cdb.cbDisplayOrder from PageTypeComposerOutputBlocks cdb inner join CollectionVersionBlocks cvb on (cdb.bID = cvb.bID and cvb.cID = cdb.cID and cvb.cvID = ?) where cdb.ptComposerFormLayoutSetControlID = ? and cdb.cID = ?';
-        $v = array($c->getVersionID(), $setControl->getPageTypeComposerFormLayoutSetControlID(), $c->getCollectionID());
-        $row = $db->GetRow($q, $v);
-        if ($row['bID'] && $row['arHandle']) {
-            $db->Execute(
+        $v = [$c->getVersionID(), $setControl->getPageTypeComposerFormLayoutSetControlID(), $c->getCollectionID()];
+        $row = $db->fetchAssoc($q, $v);
+        if ($row && $row['bID'] && $row['arHandle']) {
+            $db->executeQuery(
                 'delete from PageTypeComposerOutputBlocks where ptComposerFormLayoutSetControlID = ? and cID = ?',
-                array($setControl->getPageTypeComposerFormLayoutSetControlID(), $c->getCollectionID())
+                [$setControl->getPageTypeComposerFormLayoutSetControlID(), $c->getCollectionID()]
             );
         }
 
@@ -370,17 +376,18 @@ class BlockControl extends Control
 
     public function recordPageTypeComposerOutputBlock(\Concrete\Core\Block\Block $block)
     {
-        $db = Loader::db();
+        $app = Application::getFacadeApplication();
+        $db = $app->make('database')->connection();
         $setControl = $this->getPageTypeComposerFormLayoutSetControlObject();
-        $db->Execute(
+        $db->executeQuery(
             'insert into PageTypeComposerOutputBlocks (cID, arHandle, ptComposerFormLayoutSetControlID, cbDisplayOrder, bID) values (?, ?, ?, ?, ?)',
-            array(
+            [
                 $block->getBlockCollectionID(),
                 $block->getAreaHandle(),
                 $setControl->getPageTypeComposerFormLayoutSetControlID(),
                 $block->getBlockDisplayOrder(),
                 $block->getBlockID(),
-            )
+            ]
         );
     }
